integration: extract token_awareness.rs
