+++ scylla/tests/integration/load_balancing/mod.rs
@@ -3,3 +3,4 @@ mod lwt_optimisation;
 mod shards;
 mod simple_strategy;
 mod tablets;
+mod token_awareness;
